[lldb][test] Combine libstdc++ and libc++ vector<bool> tests into generic test
