+++ lldb/test/API/functionalities/data-formatter/data-formatter-stl/generic/vbool/Makefile
@@ -1,6 +1,4 @@
 CXX_SOURCES := main.cpp
 
-CFLAGS_EXTRAS := -O0
-USE_LIBSTDCPP := 1
-
 include Makefile.rules
+
