--- lldb/test/API/functionalities/data-formatter/data-formatter-stl/libcxx/vbool/Makefile
@@ -1,4 +1,0 @@
-CXX_SOURCES := main.cpp
-USE_LIBCPP := 1
-include Makefile.rules
-
